Remove the MD5-based manifest file system.
A long time ago, it seemed like a good idea to arrange that binaries
of my puzzles would automatically cease to identify themselves as a
particular upstream version number if any changes were made to the
source code, so that if someone made a local tweak and distributed the
result then I wouldn't get blamed for the results. Since then I've
decided the whole idea is more trouble than it's worth, so I'm
retiring it completely.


git-svn-id: svn://svn.tartarus.org/sgt/puzzles@10264 cda61777-01e9-0310-a592-d414129be87e